Compare feedback form fields. 103273.
Add an additional validation function for the feedback form that checks
to see if both fields have the same value. This should help to protect
against spam since spam submissions often use the same value in every
field, and there should never be a genuine feedback message where this
is the case.

[ Partial fix for 103273 ]