Enable the beta banner module. 105041
We use an implementation of `hook_update_N()` in the FSA report problem
module to enable the new beta banner module.

The beta banner has nothing to do with the functionality contained in
the report a problem module, it's just a convenient place to use an
update hook to avoid manual steps during deployment.

[ Partial fix for 105041 ]